Moved classes to another directory and fixed error in Client test.
